refactor(agent): move AgentRuntimeProfileModel next to the harness models
The architecture ratchet caught models.py growing past its baseline when the
runtime_type and harness_id columns were added. Rather than trim the
declaration to fit, move the model to runtime_models.py, where the
agent_host_harnesses table it now references already lives.

models.py drops from 640 to 535 lines, under the 600-line limit, so it leaves
the oversized list rather than merely stopping its growth; its stale baseline
entry is removed to lock that in.

The daemon relationship annotation becomes Mapped[Any], matching its sibling
relationships in the same class. A cross-module forward reference would not
resolve at mapper configuration time.

Verified: ruff, architecture ratchet, 2723 unit tests, and zero autogenerate
drift against a freshly migrated database.
